AutoRclone/requirements.txt
google_auth_oauthlib==0.4.0
urllib3==1.24.2
httplib2shim==0.0.3
protobuf==3.9.2
google_api_python_client==1.7.11
paramiko==2.9.2
PyMySQL==1.0.2
sshtunnel==0.4.0
psutil
filelock
psutil
filelock